+++ egui/src/introspection.rs
@@ -19,7 +19,7 @@ impl Texture {
             size *= ui.available_width() / size.x;
         }
         let (id, rect) = ui.allocate_space(size);
-        let response = ui.interact(rect, id, Sense::nothing());
+        let response = ui.interact(rect, id, Sense::hover());
         let mut triangles = Triangles::default();
         triangles.add_rect_with_uv(rect, [pos2(0.0, 0.0), pos2(1.0, 1.0)].into(), WHITE);
         ui.painter().add(PaintCmd::triangles(triangles));